Split up `video_chat.dart`
